+++ test/function/relational/larger.test.js
@@ -191,31 +191,12 @@ describe('larger', function() {
     });
   });
 
-  describe('Complex Numbers', function () {
-
-    it('should compare complex numbers', function() {
-      assert.equal(larger(complex(1,1), complex(1,1)), false);
-      assert.equal(larger(complex(2,1), complex(1,2)), true);
-      assert.equal(larger(complex(0,1), complex(1,2)), false);
-    });
-
-    it('should compare complex number and number', function() {
-      assert.equal(larger(complex(1,0), 1), false);
-      assert.equal(larger(complex(2,1), 1), true);
-      assert.equal(larger(complex(0,1), 1), false);
-      assert.equal(larger(1, complex(1,0)), false);
-      assert.equal(larger(1, complex(2,1)), false);
-      assert.equal(larger(1, complex(0,1)), true);
-    });
-
-    it('should compare complex number and bignumber', function() {
-      assert.equal(larger(complex(1,0), math.bignumber(1)), false);
-      assert.equal(larger(complex(2,1), math.bignumber(1)), true);
-      assert.equal(larger(complex(0,1), math.bignumber(1)), false);
-      assert.equal(larger(math.bignumber(1), complex(1,0)), false);
-      assert.equal(larger(math.bignumber(1), complex(2,1)), false);
-      assert.equal(larger(math.bignumber(1), complex(0,1)), true);
-    });
+  it('should throw an error when comparing complex numbers', function() {
+    assert.throws(function () {larger(complex(1,1), complex(1,2));}, TypeError);
+    assert.throws(function () {larger(complex(2,1), 3);}, TypeError);
+    assert.throws(function () {larger(3, complex(2,4));}, TypeError);
+    assert.throws(function () {larger(math.bignumber(3), complex(2,4));}, TypeError);
+    assert.throws(function () {larger(complex(2,4), math.bignumber(3));}, TypeError);
   });
 
   it('should throw an error if matrices are different sizes', function() {