+++ test/function/set/setPowerset.test.js
@@ -5,6 +5,8 @@ var math = require('../../../index');
 describe('setPowerset', function () {
   it('should return the powerset of a set', function () {
     assert.deepEqual(math.setPowerset([1, 2]), [[], [1], [2], [1, 2]]);
+    assert.deepEqual(math.setPowerset([1, math.complex(2,2)]),
+        [[], [1], [math.complex(2,2)], [1, math.complex(2,2)]]);
     assert.deepEqual(math.setPowerset([]), []);
   });