+++ test/function/set/setSize.test.js
@@ -7,6 +7,7 @@ describe('setSize', function () {
     assert.strictEqual(math.setSize([]), 0);
     assert.strictEqual(math.setSize([1]), 1);
     assert.strictEqual(math.setSize([1, 2]), 2);
+    assert.strictEqual(math.setSize([1, math.complex(2,2)]), 2);
   });
 
   it('should return the number of elements of a multiset', function () {