+++ test/function/statistics/median.test.js
@@ -68,10 +68,6 @@ describe('median', function() {
     ])), 3.5);
   });
 
-  it('should return the median for complex numbers', function() {
-    assert.deepEqual(median(new Complex(2, 3), new Complex(-1, 2)), new Complex(0.5, 2.5));
-  });
-
   it('should throw an error if called with invalid number of arguments', function() {
     assert.throws(function() {median()});
     assert.throws(function() {median([], 2, 3)});
@@ -88,6 +84,7 @@ describe('median', function() {
 
   it('should throw an error if called with unsupported type of arguments', function() {
     assert.throws(function () {median(new Date(), 2, 3)}, /TypeError: Unexpected type of argument/);
+    assert.throws(function () {median(new Complex(2,3), new Complex(-1,2))}, /TypeError: No ordering relation is defined for complex numbers/);
   });
 
   it('should throw an error if called with an empty array', function() {